git: Fixed merge conflicts
